Revert "feat: action calculations (WIP)"
This reverts commit 9c361a0f37306ecaddbb3ae5f4d9bb739fd490ec.
