+++ docs/actions/3-plot-cut-epsilon.py
@@ -5,7 +5,7 @@ from plotly.subplots import make_subplots
 from metastable.map.map import FixedPointMap, PathType
 
 # Load the fixed point map
-path = Path("/home/paul/Projects/misc/keldysh/metastable/docs/paths/examples/output/13/output_map_with_actions.npz")
+path = Path("/home/paul/Projects/misc/keldysh/metastable/docs/paths/examples/output/12/output_map_with_actions.npz")
 fixed_point_map = FixedPointMap.load(path)
 
 # Get the parameter grids
