+++ docs/paths/examples/epsilon_cut.py
@@ -59,19 +59,19 @@ if __name__ == "__main__":
     )
     fixed_point_map = FixedPointMap.load(map_path)
 
-    output_path = Path("/home/paul/Projects/misc/keldysh/metastable/docs/paths/examples/output/14")
+    output_path = Path("/home/paul/Projects/misc/keldysh/metastable/docs/paths/examples/output/13")
     
     # Create the bifurcation diagram
     fig = plot_bifurcation_diagram(fixed_point_map)
 
     # Choose a kappa index for the epsilon cut
-    kappa_idx = 60
+    kappa_idx = 40
     
     # Get the bistable epsilon range for this kappa
     epsilon_boundaries = get_bistable_epsilon_range(fixed_point_map.bistable_region, kappa_idx)
     
     # Generate epsilon sweeps
-    epsilon_sweeps = generate_sweep_index_pairs(epsilon_boundaries, sweep_fraction=0.2)
+    epsilon_sweeps = generate_sweep_index_pairs(epsilon_boundaries, sweep_fraction=0.1)
     
     # Get the actual kappa value from index
     kappa_value = fixed_point_map.kappa_linspace[kappa_idx]
